Move ListCacheKey to lists folder
