+++ packages/client/src/observable/internal/list/ListCacheKey.ts
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-import type { CollectionStorageData } from "./BaseCollectionQuery.js";
-import type { CacheKey } from "./CacheKey.js";
-import type { Canonical } from "./Canonical.js";
-import type { ListQuery } from "./ListQuery.js";
-import type { SimpleWhereClause } from "./SimpleWhereClause.js";
+import type { CollectionStorageData } from "../BaseCollectionQuery.js";
+import type { CacheKey } from "../CacheKey.js";
+import type { Canonical } from "../Canonical.js";
+import type { ListQuery } from "../ListQuery.js";
+import type { SimpleWhereClause } from "../SimpleWhereClause.js";
 
 export interface ListStorageData extends CollectionStorageData {}
